points working, user also getting 10 points on getting like on post.

+++ app/src/main/java/com/bcabuddies/letsstudy/Home/view/HomeView.java
@@ -7,5 +7,5 @@ import com.bcabuddies.letsstudy.Base.BaseView;
 public interface HomeView extends BaseView {
     void showValidationError(String message);
     void getUserDetails(Bundle user);
-    void firebaseData(String profUrl, String name, String age, String course, int points);
+    void firebaseData(String profUrl, String name, String age, String course, long points);
 }
